Replace Perlin Noise with OpenSimplex Noise
Perlin Noise, which is currently being used as the primary basefor terrain generation, is an older algorithm that tends to exhibit visually-signficiant directional artifacts. In this day and age where alternatives have been developed, it doesn't make much sense to continue to use Perlin Noise -- especially in a project such as this. This replaces the code inside PerlinNoise.java with code that generates OpenSimplex Noise instead. OpenSimplex Noise is much more visually isotropic (free of visually-significant directional artifacts).